+++ src/main/java/com/brain/pulsar/species/colonies/District.java
@@ -1,6 +1,6 @@
 package com.brain.pulsar.species.colonies;
 
-import com.brain.pulsar.buckets.ResourceBucket;
+import com.brain.pulsar.other.ResourceBucket;
 import com.brain.pulsar.other.Scope;
 import com.brain.pulsar.universe.Body;
 import com.brain.pulsar.xml.types.DistrictType;
